DPC-405: Refactor Resource setup into helper method
