+++ dpc-api/src/test/java/gov/cms/dpc/api/resources/DefinitionResourceTest.java
@@ -1,0 +1,4 @@
+package gov.cms.dpc.api.resources;
+
+public class DefinitionResourceTest {
+}
